requirements.txt
Flask
PyGithub
requests
# aiohttp==3.8.1  # Comment out or remove this line